# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "Autentica al usuario y establece una cookie HTTP-only con el JWT",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso (token en cookie)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        },
        "summary": "Iniciar sesión",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Crea una nueva cuenta de usuario",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado (token en cookie)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "409": {
            "description": "El usuario ya existe"
          }
        },
        "summary": "Registrar usuario",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Invalida el token JWT y limpia la cookie",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sesión cerrada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Cerrar sesión",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/reports/evaluation-report": {
      "post": {
        "description": "Genera un reporte completo según el formato del Paso 9 del documento, incluyendo tabla de resultados, resumen ejecutivo y datos para gráficas",
        "operationId": "ReportsController_generateEvaluationReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Datos de entrada del tratador térmico",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              },
              "examples": {
                "example1": {
                  "summary": "Ejemplo de evaluación completa",
                  "value": {
                    "diameter": 6,
                    "length": 20,
                    "totalFlow": 500,
                    "waterFraction": 20,
                    "apiGravity": 18,
                    "inletTemperature": 75,
                    "treatmentTemperature": 140,
                    "ambientTemperature": 30,
                    "operatingPressure": 50,
                    "oilRetentionTime": 60,
                    "waterRetentionTime": 30,
                    "windSpeed": 15
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reporte de evaluación generado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reportData": {
                      "type": "object"
                    },
                    "resultsTable": {
                      "type": "array"
                    },
                    "executiveSummary": {
                      "type": "object"
                    },
                    "simulationCharts": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos"
          }
        },
        "summary": "Generar reporte completo de evaluación",
        "tags": [
          "Evaluation Reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "TermoLab API",
    "description": "API para gestión de tratamientos térmicos",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "UserDataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "role": {
            "type": "string",
            "example": "user"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indica si la operación fue exitosa"
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo del resultado"
          },
          "accessToken": {
            "type": "string",
            "description": "Token de acceso JWT"
          },
          "user": {
            "$ref": "#/components/schemas/UserDataDto"
          },
          "expiresAt": {
            "type": "string",
            "example": "2023-12-31T23:59:59Z"
          }
        },
        "required": [
          "success",
          "message",
          "accessToken",
          "user",
          "expiresAt"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indica si la operación fue exitosa"
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo del resultado"
          },
          "user": {
            "description": "Datos básicos del usuario registrado",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserDataDto"
              }
            ]
          },
          "expiresAt": {
            "type": "string",
            "example": "2023-12-31T23:59:59Z",
            "description": "Fecha de expiración del token en formato ISO"
          }
        },
        "required": [
          "success",
          "message",
          "user",
          "expiresAt"
        ]
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true,
            "description": "Indica si el logout fue exitoso"
          },
          "message": {
            "type": "string",
            "example": "Sesión cerrada exitosamente",
            "description": "Mensaje descriptivo del resultado"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}